Move web wallets to folder.
