+++ client/scripts/controllers/app/webWallets/keplr_web_wallet.ts
