+++ client/scripts/controllers/app/webWallets/near_web_wallet.ts
