+++ client/scripts/controllers/chain/ethereum/marlin/adapter.ts
@@ -2,7 +2,7 @@
 import { EthereumCoin } from 'adapters/chain/ethereum/types';
 import { MPondFactory } from 'MPondFactory';
 
-import MetamaskWebWalletController from 'controllers/app/metamask_web_wallet';
+import MetamaskWebWalletController from 'controllers/app/webWallets/metamask_web_wallet';
 import EthereumAccount from 'controllers/chain/ethereum/account';
 import EthereumAccounts from 'controllers/chain/ethereum/accounts';
 import { ChainBase, ChainClass, IChainAdapter, ChainEntity, ChainEvent, NodeInfo } from 'models';
