+++ client/scripts/controllers/chain/ethereum/token/adapter.ts
@@ -3,7 +3,7 @@ import { ethers } from 'ethers';
 import { EthereumCoin } from 'adapters/chain/ethereum/types';
 
 import { Erc20Factory } from 'Erc20Factory';
-import MetamaskWebWalletController from 'controllers/app/metamask_web_wallet';
+import MetamaskWebWalletController from 'controllers/app/webWallets/metamask_web_wallet';
 import EthereumAccount from 'controllers/chain/ethereum/account';
 import EthereumAccounts from 'controllers/chain/ethereum/accounts';
 import { ChainBase, IChainAdapter, NodeInfo } from 'models';
@@ -12,7 +12,7 @@ import ChainEntityController from 'controllers/server/chain_entities';
 import { IApp } from 'state';
 
 import EthereumTokenChain from './chain';
-import ContractApi from './api';
+import TokenApi from './api';
 
 export default class Token extends IChainAdapter<EthereumCoin, EthereumAccount> {
   public readonly base = ChainBase.Ethereum;
@@ -40,7 +40,7 @@ export default class Token extends IChainAdapter<EthereumCoin, EthereumAccount> 
     await this.chain.resetApi(this.meta);
     await this.chain.initMetadata();
     await this.accounts.init(this.chain);
-    const api = new ContractApi(new Erc20Factory(), this.meta.address, this.chain.api.currentProvider as any);
+    const api = new TokenApi(new Erc20Factory(), this.meta.address, this.chain.api.currentProvider as any);
     await api.init();
     this.chain.contractApi = api;
     await super.initApi();
