+++ client/scripts/controllers/chain/ethereum/token/api.ts
@@ -2,4 +2,4 @@ import { Erc20 } from 'Erc20';
 
 import ContractApi from 'controllers/chain/ethereum/contractApi';
 
-export default class ContractApi extends ContractApi<Erc20> { }
+export default class TokenApi extends ContractApi<Erc20> { }
